Split TensorDict memmap helpers
